+++ packages/backend/src/plugins/scaffolder/actions/tyk-api-create.ts
@@ -2,7 +2,7 @@ import {Config} from '@backstage/config';
 import {Logger} from 'winston';
 import {createTemplateAction} from '@backstage/plugin-scaffolder-node';
 import {z} from 'zod';
-import {API, TykDashboardConfig} from "../../../../../../plugins/catalog-backend-module-tyk/src/clients/schemas";
+import {API, TykDashboardConfig} from "../../../../../../plugins/catalog-backend-module-tyk/src/clients/types";
 import {TykDashboardClient} from "../../../../../../plugins/catalog-backend-module-tyk/src/clients/tyk-dashboard-client";
 
 export const createTykApiAction = (config: Config, logger: Logger) => {
