+++ plugins/catalog-backend-module-tyk/src/providers/index.ts
@@ -1,0 +1,1 @@
+export { TykEntityProvider } from "./tyk-entity-provider";
