+++ plugins/catalog-backend-module-tyk/src/providers/tyk-entity-provider.ts
@@ -12,7 +12,8 @@ import {Router} from 'express';
 import {PluginTaskScheduler} from '@backstage/backend-tasks';
 import {kebabCase} from 'lodash';
 import yaml from 'js-yaml';
-import {API, TykDashboardConfig, TykConfig, TykGlobalOptionsConfig, enrichedGateway} from "../clients/schemas";
+import {API, enrichedGateway} from "../clients/types";
+import {TykDashboardConfig, TykGlobalOptionsConfig} from "./types";
 import {TykDashboardClient} from "../clients/tyk-dashboard-client";
 import { Config } from '@backstage/config';
 import { string } from 'zod';
